Patch from Scott McKellar:
1. I made __osrfChatXMLErrorOcurred and __osrfChatClientSentDisconnect
static, and removed the leading underscores from their names.  No other
source files refer to them.

2. In osrfNewChatServer() I explicitly initialize the port member to
zero -- which is already happening implicitly due to the memset() in
safe_malloc().

3. Also in osrfNewChatServer(): the existing code populates the secret
member only if the secret parameter is not NULL, and leaves it
uninitialized otherwise.  However if the secret parameter were NULL
we would have already performed an early exit.  I populate the secret
member unconditionally, just as we do with the domain member.  I also
moved this assignment up with the other assignments, for a more
readable flow.

4. In osrfChatServerFree(), the existing code leaves several memory
references dangling, without freeing them.  I added code to free the
domain member, the deadNodes member, and finally the osrfChatServer
itself.

git-svn-id: svn://svn.open-ils.org/OpenSRF/trunk@936 9efc2488-bf62-4759-914b-345cdb29e865