Patch from Scott McKellar:
1. I declared several functions as static, and moved their prototypes
from the header file to the implementation file.  In each case I
verified that nothing else in the code base calls it.

2. I removed the socket_send_nowait prototype from the header, and
commented it out in the implementation file.  I was tempted to
eliminate it entirely, but stayed my hand.  It is a trivial wrapper
for _socket_send(), and nothing calls it.

3. Nothing calls several other functions either.  However unlike
socket_send_nowait() they are non-trivial functions, and they appear
to be intended for external linkage, so I left them alone for
potential future use.

4. The identifier _socket_print_list is reserved for local scope,
and should not have external linkage, because it starts with an
underscore followed by a lower case letter.  I may submit a patch to
rename it some day if I ever get around to it, but that will require
a change in socket_test.c, which calls it.

5. I moved the definition of the RBUFSIZE macro from the header file
into the implmentation file.  It is an internal detail that no
other translation unit should know or care about.

git-svn-id: svn://svn.open-ils.org/OpenSRF/trunk@937 9efc2488-bf62-4759-914b-345cdb29e865