Patch from Scott McKellar to add stderr logging when config (and thus configured logging) is unavailable:
If osrfConfigInit() is unable to load the configuration file,
it issues error messages to the log.

Unfortunately, if we can't load the configuration file, we don't know
where the log file is.  The usual result is that the messages don't
go anywhere.  At best, depending on some convoluted details that
aren't worth going into here, the messages will go to an obscure
system-wide syslog file that only a hardened system administrator
could love.

In this patch I send similar messages to standard error, so that
someone can see them.  It might make sense to delete the calls
to osrfLogWarning(), but I left them alone.

git-svn-id: svn://svn.open-ils.org/OpenSRF/trunk@953 9efc2488-bf62-4759-914b-345cdb29e865