Patch from Ed Summers:
Attached is a minor patch to OpenSRF/src/java/Makefile for creating a
opensrf.jar. I was thinking an opensrf.jar would make it easy to
deploy OpenSRF client in a Woodchip/OpenTaps instance. It's not yet
clear to me what the build environment is (or will be) for
Woodchip--but .jar files are used to across the board to manage
external dependencies in Javaland.

- - -

I've also given the depencies their own vars for easier handling and
made the dependency fetching smarter to prevent re-fetching existing deps



git-svn-id: svn://svn.open-ils.org/OpenSRF/trunk@1052 9efc2488-bf62-4759-914b-345cdb29e865