JSON to XML patches from Scott McKellar:
Guard against multiple #inclusions.

Plug a potential memory leak in the jsonObjectToXML
function.  If the input parameter was NULL we would fail to free
the growing_buffer we had just allocated.  I rearranged it to check
for NULL before allocating the growing_buffer.

Also: I added the static qualifier to the _escape_xml function, to
match the declaration at the top of the file.



git-svn-id: svn://svn.open-ils.org/OpenSRF/trunk@1208 9efc2488-bf62-4759-914b-345cdb29e865