Patch from Scott McKellar:
1. In osrfSystemBootstrap(): we build an osrfStringArray of children
to spawn, and then spawn them.  However we didn't free the
osrfStringArray afterwards.  Now we do.

2. In osrfSystemBootstrapClientResc(): we were potentially leaking
the osrfStringArray arr in an early return.  I moved the early return
to a point before the allocation of the osrfStringArray and a number
of other things.  Not only does the early return not have to free
the osrfStringArray, it also doesn't have to free the other things
any more.



git-svn-id: svn://svn.open-ils.org/OpenSRF/trunk@1239 9efc2488-bf62-4759-914b-345cdb29e865