Patch from Scott McKellar:
In _jsonInsertParserItem() I changed a switch/case to an if/else,
eliminating a supposedly unreachable default branch that, if reached,
would leak memory.

With this change, a jsonObject that is neither a JSON_HASH nor a
JSON_ARRAY will be silently converted to a JSON_ARRAY by the call to
jsonObjectPush().



git-svn-id: svn://svn.open-ils.org/OpenSRF/trunk@1268 9efc2488-bf62-4759-914b-345cdb29e865