Patch from Scott McKellar:
1. Pedantic: I changed child_dead from an int to a sig_atomic_t, since
we set it from a signal handler.

2. In check_children() and prefork_child_wait(), we prepare an input
buffer for a read() by calling the osrf_clearbuf macro.  Depending on
the existence of NDEBUG, this macro may or may not do the right thing.
If it does the wrong thing, it will pad the input data with
exclamation points, up to the maximimum length.

I added code to explicitly add a terminal nul, thus rendering
the osrf_clearbuf macro both harmless and superfluous.

This issue is the same one that I mentioned yesterday in connection
with socket_bundle.c, so I shall not belabor it here.

3. In addition: in check_children() I arranged to issue the debug
message only if the read is successful.



git-svn-id: svn://svn.open-ils.org/OpenSRF/trunk@1278 9efc2488-bf62-4759-914b-345cdb29e865