Patch from Scott McKellar:
This patch elinminates the use of the osrf_clearbuf macro in several
spots.  In each case, the buffer needs to be initialized to an empty
string, because the call to jid_get_* doesn't necessarily populate it
with anything else.  In debugging mode, osrf_clearbuf() will fill the
field with exclamation points, which is not helpful in this context.



git-svn-id: svn://svn.open-ils.org/OpenSRF/trunk@1302 9efc2488-bf62-4759-914b-345cdb29e865