Patch from Scott McKellar:
This patch boosts the performance of an osrfHashIterator a bit, by
reusing the "current" buffer whenever possible instead of freeing and
reallocating it on every iteration.

I wrote a benchmark that repeatedly traverses a 27-item hash table.
The new version is about 10% faster.  This improvement is not as
dramatic as I had hoped for, but it's easy to get, and it won't hurt.



git-svn-id: svn://svn.open-ils.org/OpenSRF/trunk@1305 9efc2488-bf62-4759-914b-345cdb29e865