Patch from Scott McKellar:
This patch tweaks the jid_get_* functions, which copy various
fragments of an jabber id into a buffer supplied by the caller.

If these functions don't find the fragments they're looking for, they
now return an empty string in the buffer.  Earlier they would leave
the buffer unchanged, leaving it to the calling code to initialize
the buffer.



git-svn-id: svn://svn.open-ils.org/OpenSRF/trunk@1321 9efc2488-bf62-4759-914b-345cdb29e865