Patch from Scott McKellar:
This patch tweaks a few things.

1. In json_parse_json_string() we declare a character array buff[],
fill it with nuls, and never refer to it again.  I eliminated it.

2. A few lines below that, we use memset() to clear a three-character
buffer.  I replaced the memset() with an initializer clause.

3. in json_handle_error() we were relying on the osrf_clearbuf macro
to add a terminal nul.  However the plan is for this macro to become
a no-op someday.  I added the nul manually.



git-svn-id: svn://svn.open-ils.org/OpenSRF/trunk@1322 9efc2488-bf62-4759-914b-345cdb29e865