Patch from Scott McKellar:
These patches provide a new and more efficient implementation of
osrfHash, using a hash table for random lookups and a doubly linked
list for iterations.

It is more efficient for two main reasons.  First, for iterations
it doesn't maintain a separate copy of the keys, that then have to go 
through the hashing algorithm.  Second, when updating an existing
item, it updates it in place rather than deleting it and creating 
a new one.



git-svn-id: svn://svn.open-ils.org/OpenSRF/trunk@1323 9efc2488-bf62-4759-914b-345cdb29e865