Patch from Kevin Beswick to use autotools for OpenSRF configuration and build
Included is a patch for a working implementation of the GNU Autotools
to OpenSRF. Some of the perks of this system are: the user is given
the ability to install OpenSRF to non-standard locations,
dependency/library checking, gives a user the ability to only install
certain components, more portable to other operating systems, etc.

To test, simply run:
autogen.sh
configure
make
make install

Options for configure can be found using:
./configure --help

Also, the patch contains a new file, osrf_config, which is generated
by running the configure script. Running this file will print all
build options a user has chosen when compiling/installing Opensrf.
Command line arguments can be added to return each single build option
as well. This is useful to developers in that they are able to locate
a certain installation directory on a non-standard installation of
OpenSRF. To see the available command line options, run: osrf_config --help



git-svn-id: svn://svn.open-ils.org/OpenSRF/trunk@1348 9efc2488-bf62-4759-914b-345cdb29e865